% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSAS.R
\name{dsas}
\alias{dsas}
\title{sinh arcshinh Probability Density Function}
\usage{
dsas(x, delta, log = FALSE)
}
\arguments{
\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{x:}{vector of quantiles.}

\item{delta:}{shape parameter.}
}
\value{
density of the sinh-arcsinh distribution with shape parameter delta
}
\description{
sinh arcshinh Probability Density Function
}
