% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSAS.R
\name{dtpsas}
\alias{dtpsas}
\title{twopiece sinh arcshinh Probability Density Function}
\usage{
dtpsas(x, mu, par1, par2, delta, param = "tp", log = FALSE)
}
\arguments{
\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{x:}{vector of quantiles.}

\item{delta:}{shape parameter.}

\item{par1:}{scale parameter 1}

\item{par2:}{scale parameter 2 or skewness parameter}

\item{param:}{parameterisation. Defaults to "tp". Information about the "eps" and "isf" parameterizations can be found in the References.}
}
\value{
density of the twopiece sinh-arcsinh distribution
}
\description{
twopiece sinh arcshinh Probability Density Function
}
