% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSAS.R
\name{psas}
\alias{psas}
\title{sinh-arcsinh Cumulative Distribution Function}
\usage{
psas(x, delta, log.p = FALSE)
}
\arguments{
\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{x:}{vector of quantiles.}

\item{delta:}{shape parameter.}
}
\value{
cumulative distribution of the twopiece sinh-arcsinh distribution
}
\description{
sinh-arcsinh Cumulative Distribution Function
}
