% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSAS.R
\name{ptpsas}
\alias{ptpsas}
\title{twopiece sinh-arcsinh Cumulative Distribution Function}
\usage{
ptpsas(x, mu, par1, par2, delta, param = "tp", log.p = FALSE)
}
\arguments{
\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{x:}{vector of quantiles.}

\item{delta:}{shape parameter.}

\item{par1:}{scale parameter 1}

\item{par2:}{scale parameter 2 or skewness parameter}

\item{param:}{parameterisation. Defaults to "tp". Information about the "eps" and "isf" parameterizations can be found in the References.}
}
\value{
cumulative distribution of the twopiece sinh-arcsinh distribution with shape parameter delta
}
\description{
twopiece sinh-arcsinh Cumulative Distribution Function
}
