% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSAS.R
\name{qsas}
\alias{qsas}
\title{sinh-arcsinh Quantile Function}
\usage{
qsas(p, delta)
}
\arguments{
\item{p:}{vector of probabilities.}

\item{delta:}{shape parameter.}
}
\value{
quantile function of the sinh-arcsinh distribution with shape parameter delta
}
\description{
sinh-arcsinh Quantile Function
}
