% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSAS.R
\name{qtpsas}
\alias{qtpsas}
\title{twopiece sinh-arcsinh Quantile Function}
\usage{
qtpsas(p, mu, par1, par2, delta, param = "tp")
}
\arguments{
\item{p:}{vector of probabilities}

\item{delta:}{shape parameter.}

\item{par1:}{scale parameter 1}

\item{par2:}{scale parameter 2 or skewness parameter}

\item{param:}{parameterisation. Defaults to "tp". Information about the "eps" and "isf" parameterizations can be found in the References.}
}
\value{
quantile function of the twopiece sinh-arcsinh distribution
}
\description{
twopiece sinh-arcsinh Quantile Function
}
